+++ penman/layout.py
@@ -280,7 +280,9 @@ def configure(g: Graph,
         data_count = len(data)
         if var is None or data_count == 0:
             raise LayoutError('possibly disconnected graph')
+
         _, surprising = _configure_node(var, data, nodemap, model)
+
         if len(data) == data_count and surprising:
             skipped.insert(0, data.pop())
         elif len(data) >= data_count:
@@ -288,13 +290,17 @@ def configure(g: Graph,
         else:
             data = skipped + data
             skipped.clear()
+
         # remove any superfluous POPs
         while data and isinstance(data[-1], Pop):
             data.pop()
     if skipped:
         raise LayoutError('incomplete configuration')
+
+    _process_epigraph(node)
     tree = Tree(node, metadata=g.metadata)
     logger.debug('Configured: %s', tree)
+
     return tree
 
 
@@ -312,13 +318,13 @@ def _configure(g, top, model):
         raise LayoutError(f'top is not a variable: {top!r}')
     nodemap[top] = (top, [])
 
-    data = list(reversed(_preconfigure(g)))
+    data = list(reversed(_preconfigure(g, model)))
     node, _ = _configure_node(top, data, nodemap, model)
 
     return node, data, nodemap
 
 
-def _preconfigure(g):
+def _preconfigure(g, model):
     """
     Arrange the triples and epidata for ordered traversal.
 
@@ -327,46 +333,37 @@ def _preconfigure(g):
     data = []
     epidata = g.epidata
     pushed = set()
+
     for triple in g.triples:
-        triple, push, pops = _preconfigure_triple(triple, pushed, epidata)
-        data.append((triple, push))
+        var, role, target = triple
+        epis, push, pops = [], False, []
+
+        for epi in epidata.get(triple, []):
+            if isinstance(epi, Push):
+                pvar = epi.variable
+                if pvar in pushed:
+                    logger.warning(
+                        f"ignoring secondary node contexts for '{pvar}'"
+                    )
+                    continue  # change to 'pass' to allow multiple contexts
+                if pvar not in (var, target) or role == CONCEPT_ROLE:
+                    logger.warning(
+                        f"node context '{pvar}' invalid for triple: {triple!r}"
+                    )
+                    continue
+                if pvar == var:
+                    triple = model.invert(triple)
+                pushed.add(pvar)
+                push = True
+            elif isinstance(epi, Pop):
+                pops.append(epi)
+            else:
+                epis.append(epi)
+
+        data.append((triple, push, epis))
         data.extend(pops)
+
     return data
-
-
-def _preconfigure_triple(triple, pushed, epidata):
-    """
-    Validate layout by markers and format other markers onto role or target.
-    """
-    var, role, target = triple
-    push, pops = None, []
-
-    for epi in epidata.get(triple, []):
-        if isinstance(epi, Push):
-            pvar = epi.variable
-            if push is not None or pvar in pushed:
-                logger.warning(
-                    f"ignoring secondary node contexts for '{pvar}'")
-                continue  # change to 'pass' to allow multiple contexts
-            if pvar not in (var, target) or role == CONCEPT_ROLE:
-                logger.warning(
-                    f"node context '{pvar}' invalid for triple: {triple!r}")
-                continue
-            pushed.add(pvar)
-            push = epi
-        elif isinstance(epi, Pop):
-            pops.append(epi)
-        elif epi.mode == 1:  # role epidata
-            role = f'{role!s}{epi!s}'
-        elif target and epi.mode == 2:  # target epidata
-            target = f'{target!s}{epi!s}'
-        else:
-            logger.warning('epigraphical marker ignored: %r', epi)
-
-    if push and pops:
-        logger.warning(
-            f'incompatible node context changes on triple: {triple!r}')
-    return (var, role, target), push, pops
 
 
 def _configure_node(var, data, nodemap, model):
@@ -387,39 +384,34 @@ def _configure_node(var, data, nodemap, model):
         datum = data.pop()
         if isinstance(datum, Pop):
             break
-
-        triple, push = datum
+        triple, push, epis = datum
+        # Finalize triple orientation
         if triple[0] == var:
-            source, role, target = triple
+            _, role, target = triple  # expected situation
         elif triple[2] == var and triple[1] != CONCEPT_ROLE:
-            target, role, source = triple
-            role, _, alignment = role.partition('~')
-            role = f'{model.invert_role(role)}{_}{alignment}'
+            _, role, target = model.invert(triple)  # unexpected inversion
+            push = False  # preconfigured push site may no longer be valid
+            surprising = True
         else:
-            # misplaced triple
-            data.append(datum)
+            data.append(datum)  # cannot place triple
             surprising = True
             break
 
+        # Insert into tree, recursively configuring nodes
         if role == CONCEPT_ROLE:
             if not target:
                 continue  # prefer (a) over (a /) when concept is missing
-            role = '/'
-            index = 0
-            push = None  # never push on concept roles
+            edges.insert(0, ('/', target, epis))
         else:
-            index = len(edges)
-
-        if push and push.variable == target:
-            nodemap[push.variable] = (push.variable, [])
-            target, _surprising = _configure_node(
-                push.variable, data, nodemap, model)
-            surprising &= _surprising
-        elif role != '/' and target in nodemap and nodemap[target] is None:
-            # site of potential node context
-            nodemap[target] = node
-
-        edges.insert(index, (role, target))
+            if push:
+                nodemap[target] = (target, [])
+                target, _surprising = _configure_node(
+                    target, data, nodemap, model
+                )
+                surprising &= _surprising
+            elif target in nodemap and nodemap[target] is None:
+                nodemap[target] = node  # site of potential node context
+            edges.append((role, target, epis))
 
     return node, surprising
 
@@ -467,6 +459,23 @@ def _get_or_establish_site(var, nodemap):
         return True
     # var is not yet available
     return False
+
+
+def _process_epigraph(node):
+    """Format epigraph data onto roles and targets."""
+    _, edges = node
+    for i, (role, target, epis) in enumerate(edges):
+        atomic_target = is_atomic(target)
+        for epi in epis:
+            if epi.mode == 1:  # role epidata
+                role = f'{role!s}{epi!s}'
+            elif epi.mode == 2 and atomic_target:  # target epidata
+                target = f'{target!s}{epi!s}'
+            else:
+                logger.warning('epigraphical marker ignored: %r', epi)
+        if not atomic_target:
+            _process_epigraph(target)
+        edges[i] = (role, target)
 
 
 def reconfigure(g: Graph,